+++ Src/Metrics.Tests/Core/DefaultContextTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using FluentAssertions;
+using Metrics.MetricData;
 using Xunit;
 
 namespace Metrics.Tests.Core