+++ Src/Metrics.Tests/Sampling/ExponentiallyDecayingReservoirTests.cs
@@ -8,6 +8,14 @@ namespace Metrics.Tests.Sampling
 {
     public class ExponentiallyDecayingReservoirTests
     {
+        private readonly TestClock clock = new TestClock();
+        private readonly TestScheduler scheduler;
+
+        public ExponentiallyDecayingReservoirTests()
+        {
+            this.scheduler = new TestScheduler(clock);
+        }
+
         [Fact]
         public void EDR_ReservoirOf100OutOf1000Elements()
         {
@@ -18,7 +26,7 @@ namespace Metrics.Tests.Sampling
             }
 
             reservoir.Size.Should().Be(100);
-            var snapshot = reservoir.Snapshot;
+            var snapshot = reservoir.GetSnapshot();
             snapshot.Size.Should().Be(100);
             snapshot.Values.Should().OnlyContain(v => 0 <= v && v < 1000);
         }
@@ -33,7 +41,7 @@ namespace Metrics.Tests.Sampling
             }
 
             reservoir.Size.Should().Be(10);
-            var snapshot = reservoir.Snapshot;
+            var snapshot = reservoir.GetSnapshot();
             snapshot.Size.Should().Be(10);
             snapshot.Values.Should().OnlyContain(v => 0 <= v && v < 10);
         }
@@ -48,7 +56,7 @@ namespace Metrics.Tests.Sampling
             }
 
             reservoir.Size.Should().Be(100);
-            var snapshot = reservoir.Snapshot;
+            var snapshot = reservoir.GetSnapshot();
             snapshot.Size.Should().Be(100);
             snapshot.Values.Should().OnlyContain(v => 0 <= v && v < 100);
         }
@@ -56,9 +64,6 @@ namespace Metrics.Tests.Sampling
         [Fact]
         public void EDR_longPeriodsOfInactivityShouldNotCorruptSamplingState()
         {
-            TestClock clock = new TestClock();
-            TestScheduler scheduler = new TestScheduler(clock);
-
             ExponentiallyDecayingReservoir reservoir = new ExponentiallyDecayingReservoir(10, 0.015, clock, scheduler);
 
             // add 1000 values at a rate of 10 values/second
@@ -68,8 +73,8 @@ namespace Metrics.Tests.Sampling
                 clock.Advance(TimeUnit.Milliseconds, 100);
             }
 
-            reservoir.Snapshot.Size.Should().Be(10);
-            reservoir.Snapshot.Values.Should().OnlyContain(v => 1000 <= v && v < 2000);
+            reservoir.GetSnapshot().Size.Should().Be(10);
+            reservoir.GetSnapshot().Values.Should().OnlyContain(v => 1000 <= v && v < 2000);
 
             // wait for 15 hours and add another value.
             // this should trigger a rescale. Note that the number of samples will be reduced to 2
@@ -77,7 +82,7 @@ namespace Metrics.Tests.Sampling
             // zero after rescale.
             clock.Advance(TimeUnit.Hours, 15);
             reservoir.Update(2000);
-            var snapshot = reservoir.Snapshot;
+            var snapshot = reservoir.GetSnapshot();
             snapshot.Size.Should().Be(2);
             snapshot.Values.Should().OnlyContain(v => 1000 <= v && v < 3000);
 
@@ -88,7 +93,7 @@ namespace Metrics.Tests.Sampling
                 clock.Advance(TimeUnit.Milliseconds, 100);
             }
 
-            var finalSnapshot = reservoir.Snapshot;
+            var finalSnapshot = reservoir.GetSnapshot();
 
             finalSnapshot.Size.Should().Be(10);
             // TODO: double check the Skip first value - sometimes first value is 2000 - which might or not be correct
@@ -98,9 +103,7 @@ namespace Metrics.Tests.Sampling
         [Fact]
         public void EDR_SpotLift()
         {
-            TestClock clock = new TestClock();
-            TestScheduler scheduler = new TestScheduler(clock);
-            ExponentiallyDecayingReservoir reservoir = new ExponentiallyDecayingReservoir(1000, 0.015, clock, scheduler);
+            ExponentiallyDecayingReservoir reservoir = new ExponentiallyDecayingReservoir(clock, scheduler);
 
             int valuesRatePerMinute = 10;
             int valuesIntervalMillis = (int)(TimeUnit.Minutes.ToMilliseconds(1) / valuesRatePerMinute);
@@ -119,15 +122,13 @@ namespace Metrics.Tests.Sampling
             }
 
             // expect that quantiles should be more about mode 2 after 10 minutes
-            reservoir.Snapshot.Median.Should().Be(9999);
+            reservoir.GetSnapshot().Median.Should().Be(9999);
         }
 
         [Fact]
         public void EDR_SpotFall()
         {
-            TestClock clock = new TestClock();
-            TestScheduler scheduler = new TestScheduler(clock);
-            ExponentiallyDecayingReservoir reservoir = new ExponentiallyDecayingReservoir(1000, 0.015, clock, scheduler);
+            ExponentiallyDecayingReservoir reservoir = new ExponentiallyDecayingReservoir(clock, scheduler);
 
             int valuesRatePerMinute = 10;
             int valuesIntervalMillis = (int)(TimeUnit.Minutes.ToMilliseconds(1) / valuesRatePerMinute);
@@ -146,15 +147,13 @@ namespace Metrics.Tests.Sampling
             }
 
             // expect that quantiles should be more about mode 2 after 10 minutes
-            reservoir.Snapshot.Percentile95.Should().Be(178);
+            reservoir.GetSnapshot().Percentile95.Should().Be(178);
         }
 
         [Fact]
         public void EDR_QuantiliesShouldBeBasedOnWeights()
         {
-            TestClock clock = new TestClock();
-            TestScheduler scheduler = new TestScheduler(clock);
-            ExponentiallyDecayingReservoir reservoir = new ExponentiallyDecayingReservoir(1000, 0.015, clock, scheduler);
+            ExponentiallyDecayingReservoir reservoir = new ExponentiallyDecayingReservoir(clock, scheduler);
 
             for (int i = 0; i < 40; i++)
             {
@@ -168,27 +167,25 @@ namespace Metrics.Tests.Sampling
                 reservoir.Update(9999);
             }
 
-            reservoir.Snapshot.Size.Should().Be(50);
+            reservoir.GetSnapshot().Size.Should().Be(50);
 
             // the first added 40 items (177) have weights 1 
             // the next added 10 items (9999) have weights ~6 
             // so, it's 40 vs 60 distribution, not 40 vs 10
-            reservoir.Snapshot.Median.Should().Be(9999);
-            reservoir.Snapshot.Percentile75.Should().Be(9999);
+            reservoir.GetSnapshot().Median.Should().Be(9999);
+            reservoir.GetSnapshot().Percentile75.Should().Be(9999);
         }
 
         [Fact]
         public void EDR_RecordsUserValue()
         {
-            TestClock clock = new TestClock();
-            TestScheduler scheduler = new TestScheduler(clock);
-            ExponentiallyDecayingReservoir reservoir = new ExponentiallyDecayingReservoir(1000, 0.015, clock, scheduler);
+            ExponentiallyDecayingReservoir reservoir = new ExponentiallyDecayingReservoir(clock, scheduler);
 
             reservoir.Update(2L, "B");
             reservoir.Update(1L, "A");
 
-            reservoir.Snapshot.MinUserValue.Should().Be("A");
-            reservoir.Snapshot.MaxUserValue.Should().Be("B");
+            reservoir.GetSnapshot().MinUserValue.Should().Be("A");
+            reservoir.GetSnapshot().MaxUserValue.Should().Be("B");
         }
     }
 }