+++ Src/Metrics/Sampling/UniformReservoir.cs
@@ -22,6 +22,8 @@ namespace Metrics.Sampling
             this.values = new UserValueWrapper[size];
         }
 
+        public long Count { get { return this.count.Value; } }
+
         public int Size
         {
             get
@@ -30,23 +32,26 @@ namespace Metrics.Sampling
             }
         }
 
-        public Snapshot Snapshot
+        public Snapshot GetSnapshot(bool resetReservoir = false)
         {
-            get
+            var size = this.Size;
+            if (size == 0)
             {
-                var size = this.Size;
-                if (size == 0)
-                {
-                    return new UniformSnapshot(Enumerable.Empty<long>());
-                }
-
-                UserValueWrapper[] values = new UserValueWrapper[size];
-                Array.Copy(this.values, values, size);
-                Array.Sort(values, UserValueWrapper.Comparer);
-                var minValue = values[0].UserValue;
-                var maxValue = values[size - 1].UserValue;
-                return new UniformSnapshot(values.Select(v => v.Value), valuesAreSorted: true, minUserValue: minValue, maxUserValue: maxValue);
+                return new UniformSnapshot(0, Enumerable.Empty<long>());
             }
+
+            UserValueWrapper[] values = new UserValueWrapper[size];
+            Array.Copy(this.values, values, size);
+
+            if (resetReservoir)
+            {
+                count.SetValue(0L);
+            }
+
+            Array.Sort(values, UserValueWrapper.Comparer);
+            var minValue = values[0].UserValue;
+            var maxValue = values[size - 1].UserValue;
+            return new UniformSnapshot(this.count.Value, values.Select(v => v.Value), valuesAreSorted: true, minUserValue: minValue, maxUserValue: maxValue);
         }
 
         public void Update(long value, string userValue = null)