+++ Src/Metrics/Sampling/WeightedSnapshot.cs
@@ -18,8 +18,9 @@ namespace Metrics.Sampling
         }
     }
 
-    public struct WeightedSnapshot : Snapshot
+    public sealed class WeightedSnapshot : Snapshot
     {
+        private readonly long count;
         private readonly long[] values;
         private readonly double[] normWeights;
         private readonly double[] quantiles;
@@ -37,8 +38,9 @@ namespace Metrics.Sampling
             }
         }
 
-        public WeightedSnapshot(IEnumerable<WeightedSample> values)
+        public WeightedSnapshot(long count, IEnumerable<WeightedSample> values)
         {
+            this.count = count;
             var sample = values.ToArray();
             Array.Sort(sample, WeightedSampleComparer.Instance);
 
@@ -62,6 +64,7 @@ namespace Metrics.Sampling
             this.maxUserValue = sample.Select(s => s.UserValue).LastOrDefault();
         }
 
+        public long Count { get { return this.count; } }
         public int Size { get { return this.values.Length; } }
 
         public long Max { get { return this.values.LastOrDefault(); } }
@@ -121,7 +124,7 @@ namespace Metrics.Sampling
 
         public double GetValue(double quantile)
         {
-            if (quantile < 0.0 || quantile > 1.0)
+            if (quantile < 0.0 || quantile > 1.0 || double.IsNaN(quantile))
             {
                 throw new ArgumentException(string.Format("{0} is not in [0..1]", quantile));
             }